Section 26: Authorization, Policies, Gates

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,7 @@ use App\Models\BlogPost;
 use Illuminate\Auth\Events\Validated;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -35,6 +36,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        // $this->authorize('create', BlogPost::class);
         return view('posts.create');
     }
 
@@ -72,7 +74,9 @@ class PostController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+        $post = BlogPost::findOrFail($id);
+        $this->authorize('update', $post);
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -85,6 +89,13 @@ class PostController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't update this post.");
+        // }
+
+        $this->authorize('update', $post);
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -101,6 +112,13 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('delete-post', $post)) {
+        //     abort(403, "You can't delete this post.");
+        // }
+
+        $this->authorize('delete', $post);
+
         $post->delete();
         session()->flash('status', 'Blog post was deleted!');
         return redirect()->route('posts.index');
